[feat/#58] textfield 입력 조건 설정
